pracovanie so senzorom, pamatanie si premennych pri otoceni obrazovky

+++ app/src/main/java/com/example/watertracker/screens/HomeScreen.kt
@@ -28,6 +28,7 @@ import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
+import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
@@ -35,7 +36,7 @@ import com.example.watertracker.R
 import kotlinx.coroutines.delay
 
 @Composable
-fun HomeScreen(totalWater: Int) {
+fun HomeScreen(totalWater: Float) {
 
     val goal = 2500f
     val progress = (totalWater / goal).coerceIn(0f, 2f)
